Break up event
